+++ src/grid.h
@@ -1,12 +1,14 @@
 #ifndef GRID_H
 #define GRID_H
 
+#include <stdlib.h>
+#include <stdint.h>
 #include "point.h"
 #include "snake.h"
 
 typedef struct Grid2D {
-    unsigned short row_size;
-    unsigned short col_size;
+    uint8_t row_size;
+    uint8_t col_size;
 
     Point food;
     Snake* snake;
@@ -16,7 +18,7 @@ static const char GRID_EMPTY_CELL = ' ';
 static const char GRID_FOOD_CELL = 'F';
 static const char GRID_SNAKE_CELL = 'S';
 
-Grid* init_grid(unsigned short, unsigned short);
+Grid* init_grid(uint8_t, uint8_t);
 void free_grid(Grid*);
 PointQueue* grid_empty_cells(Grid*);
 Point grid_spawn_food(Grid*);
